pyproject.toml
[project]
name = "llm-eval-framework"
version = "0.1.0"
description = "Multi-dimensional LLM eval framework for a telecom NOC diagnostic agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.89.0",
    "langchain-anthropic>=1.4.0",
    "langgraph>=1.1.6",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that call the Anthropic API (skipped by default, run with -m integration)",
]
# Unit tests only by default; add -m integration to include API tests
addopts = "-m 'not integration' -v"
